+++ packages/expo-auth-session/src/SessionUrlProvider.ts
@@ -25,11 +25,7 @@ export class SessionUrlProvider {
     });
   }
 
-  getStartUrl(
-    authUrl: string,
-    returnUrl: string,
-    proxyProjectIdOverride: string | undefined
-  ): string {
+  getStartUrl(authUrl: string, returnUrl: string, projectNameForProxy: string | undefined): string {
     if (Platform.OS === 'web' && !Platform.isDOMAvailable) {
       // Return nothing in SSR envs
       return '';
@@ -39,10 +35,10 @@ export class SessionUrlProvider {
       returnUrl,
     });
 
-    return `${this.getRedirectUrl({ proxyProjectIdOverride })}/start?${queryString}`;
+    return `${this.getRedirectUrl({ projectNameForProxy })}/start?${queryString}`;
   }
 
-  getRedirectUrl(options: { proxyProjectIdOverride?: string; urlPath?: string }): string {
+  getRedirectUrl(options: { projectNameForProxy?: string; urlPath?: string }): string {
     if (Platform.OS === 'web') {
       if (Platform.isDOMAvailable) {
         return [window.location.origin, options.urlPath].filter(Boolean).join('/');
@@ -52,13 +48,13 @@ export class SessionUrlProvider {
       }
     }
 
-    const legacyExpoProjectId =
-      options.proxyProjectIdOverride ||
+    const legacyExpoProjectFullName =
+      options.projectNameForProxy ||
       Constants.manifest?.originalFullName ||
       Constants.manifest2?.extra?.expoClient?.originalFullName ||
       Constants.manifest?.id;
 
-    if (!legacyExpoProjectId) {
+    if (!legacyExpoProjectFullName) {
       let nextSteps = '';
       if (__DEV__) {
         if (Constants.executionEnvironment === ExecutionEnvironment.Bare) {
@@ -72,18 +68,18 @@ export class SessionUrlProvider {
 
       if (Constants.manifest2) {
         nextSteps =
-          ' The AuthSession proxy has been deprecated for modern applications. Prefer AuthRequest (with the useProxy option set to false) in combination with an Expo Development Client build of your application.' +
-          ' To continue using the AuthSession proxy during the transition, specify the project ID using the proxyProjectIdOverride option.';
+          ' Prefer AuthRequest (with the useProxy option set to false) in combination with an Expo Development Client build of your application.' +
+          ' To continue using the AuthSession proxy, specify the project full name (@owner/slug) using the projectNameForProxy option.';
       }
 
       throw new Error(
-        'Cannot use AuthSession proxy because the project ID is not defined.' + nextSteps
+        'Cannot use the AuthSession proxy because the project full name is not defined.' + nextSteps
       );
     }
 
-    const redirectUrl = `${SessionUrlProvider.BASE_URL}/${legacyExpoProjectId}`;
+    const redirectUrl = `${SessionUrlProvider.BASE_URL}/${legacyExpoProjectFullName}`;
     if (__DEV__) {
-      SessionUrlProvider.warnIfAnonymous(legacyExpoProjectId, redirectUrl);
+      SessionUrlProvider.warnIfAnonymous(legacyExpoProjectFullName, redirectUrl);
       // TODO: Verify with the dev server that the manifest is up to date.
     }
     return redirectUrl;